# services/gateway/docs/swagger.yaml
basePath: /
definitions:
  ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  presenceResponse:
    properties:
      online:
        type: boolean
      user_id:
        type: string
    type: object
...
  renditionResponse:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      duration:
        type: number
      height:
        type: integer
      id:
        type: string
      size:
        type: integer
    type: object
...
  videoAlbumListItem:
    properties:
      id:
        type: string
      title:
        type: string
      videos:
        items:
          $ref: '#/definitions/videoResponse'
        type: array
    type: object
  videoAlbumListResponse:
    properties:
      albums:
        items:
          $ref: '#/definitions/videoAlbumListItem'
        type: array
    type: object
  videoAlbumResponse:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
  videoFlagResponse:
    properties:
      bookmarked:
        type: boolean
      favorited:
        type: boolean
    type: object
  videoLikeResponse:
    properties:
      liked:
        type: boolean
      likes_count:
        type: integer
    type: object
  videoResponse:
    properties:
      album_id:
        type: string
      created_at:
        type: string
      duration:
        type: number
      id:
        type: string
      is_bookmarked_by_current:
        type: boolean
      is_favorited_by_current:
        type: boolean
      is_liked_by_current:
        type: boolean
      likes_count:
        type: integer
      media_id:
        type: string
      original_filename:
        type: string
      owner_id:
        type: string
      renditions:
        items:
          $ref: '#/definitions/renditionResponse'
        type: array
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      views_count:
        type: integer
    type: object
  videoViewResponse:
    properties:
      views_count:
        type: integer
    type: object
info:
  contact: {}
  description: Единый публичный HTTP-контракт Go-бэкенда.
  title: General Project API
  version: "1.0"
paths:
...
  /v1/media/videos:
    get:
      parameters:
      - description: UUID владельца
        in: query
        name: owner_id
        type: string
      - description: uploaded, favorite, viewed или bookmarked
        in: query
        name: tab
        type: string
      - description: Поиск по названию
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/videoAlbumListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить видео по альбомам
      tags:
      - media
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Видео
        in: formData
        name: file
        required: true
        type: file
      - description: Название видео
        in: formData
        name: title
        type: string
      - description: UUID альбома
        in: formData
        name: album_id
        type: string
      - description: Высоты через запятую
        in: formData
        name: heights
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/videoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Загрузить видео
      tags:
      - media
  /v1/media/videos/{videoID}:
    delete:
      parameters:
      - description: UUID видео
        in: path
        name: videoID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Удалить видео
      tags:
      - media
    get:
      parameters:
      - description: UUID видео
        in: path
        name: videoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/videoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить видео
      tags:
      - media
  /v1/media/videos/{videoID}/bookmark:
    delete:
      parameters:
      - description: UUID видео
        in: path
        name: videoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/videoFlagResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Изменить закладку видео
      tags:
      - media
    post:
      parameters:
      - description: UUID видео
        in: path
        name: videoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/videoFlagResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Изменить закладку видео
      tags:
      - media
  /v1/media/videos/{videoID}/favorite:
    delete:
      parameters:
      - description: UUID видео
        in: path
        name: videoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/videoFlagResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Изменить избранное видео
      tags:
      - media
    post:
      parameters:
      - description: UUID видео
        in: path
        name: videoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/videoFlagResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Изменить избранное видео
      tags:
      - media
  /v1/media/videos/{videoID}/like:
    post:
      parameters:
      - description: UUID видео
        in: path
        name: videoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/videoLikeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Поставить или убрать like видео
      tags:
      - media
  /v1/media/videos/{videoID}/view:
    post:
      parameters:
      - description: UUID видео
        in: path
        name: videoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/videoViewResponse'
      summary: Зафиксировать просмотр видео
      tags:
      - media
  /v1/media/videos/albums:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Название альбома
        in: formData
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/videoAlbumResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Создать альбом видео
      tags:
      - media
  /v1/media/videos/albums/{albumID}:
    delete:
      parameters:
      - description: UUID альбома
        in: path
        name: albumID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Удалить альбом видео
      tags:
      - media
...
  /v1/messaging/presence/{userID}:
    get:
      parameters:
      - description: UUID пользователя
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Проверить presence пользователя
      tags:
      - messaging
  /v1/notifications/stream:
    get:
      parameters:
      - description: Access-токен для EventSource
        in: query
        name: access_token
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Открыть поток уведомлений
      tags:
      - messaging
...
swagger: "2.0"